Handle xterm copy-to-clipboard feature for Tmux
This allows using Tmux default setting for terminals
that identify themselves as xterm* to transparently
receive copies of the current selection.

With proper configuration, you can now yank text in Vim
running inside of Tmux and immediately paste to a Windows app.

Currently the copy-buffer size limit is a 64k.